requirements.txt
fastapi
uvicorn
pydantic
slowapi
python-dotenv
python-jose[cryptography]
bcrypt==4.0.0
passlib[bcrypt]
typing-extensions
pydantic-ai-slim[duckduckgo]
pydantic-ai-slim[a2a]
weasyprint
redis
celery
jinja2
sqlalchemy
pydantic_ai_slim[duckduckgo]
psycopg2
faker